+++ modules/mariadb/test/test_mariadb_runner.ml
@@ -1,0 +1,4 @@
+let () =
+  Lwt_main.run
+    (Alcotest_lwt.run "testcontainers-mariadb"
+       [ ("mariadb", Test_mariadb.suite) ])